Implementado librerias, hay que implementar la propia UART y revisar lo demás
