+++ Code/TP3-CDyM/Debug/makedep.mk
@@ -1,0 +1,14 @@
+################################################################################
+# Automatically-generated file. Do not edit or delete the file
+################################################################################
+
+main.c
+
+ringtone.c
+
+serialPort.c
+
+timer.c
+
+uart.c
+
