+++ Code/TP3-CDyM/main.h
@@ -1,0 +1,21 @@
+/*
+ * main.h
+ *
+ * Created: 5/6/2023 15:06:47
+ *  Author: Infinio
+ */ 
+
+
+#ifndef MAIN_H_
+#define MAIN_H_
+
+#include "stdio.h"
+#include "stdlib.h"
+#include "stdint.h"
+#include <avr/io.h>
+#include "avr/interrupt.h"
+#define F_CPU 16000000UL
+
+
+
+#endif /* MAIN_H_ */
