+++ Code/TP3-CDyM/ringtone.h
@@ -1,0 +1,55 @@
+/*
+ * ringtone.h
+ *
+ * Created: 5/6/2023 14:49:41
+ *  Author: Infinio
+ */ 
+
+
+#ifndef RINGTONE_H_
+#define RINGTONE_H_
+
+#include "stdio.h"
+#include "stdlib.h"
+#include "stdint.h"
+#include <avr/io.h>
+#include "avr/interrupt.h"
+#define F_CPU 16000000UL
+
+// Colecci?n de m?sica RTTL
+const char *rtttl_library[]=
+{
+	"SeaShanty2:d=16,o=5,b=180:5a,p,5e,5d,5c#,p,5c#,5d,5e,5f#,5g#,5e,p,5f#,p,5e,5d,5c#,p,5c#,4h,p,5c#,p,5a,p,5e,5d,5c#,p,5c#,5d,5e,5f#,5d,5f#,5e,5d,5c#,4h,5c#,5d,5f#,5e,5d,5c#,4h,4a",
+	
+	"TheSimpsons:d=4,o=5,b=160:c.6,e6,f#6,8a6,g.6,e6,c6,8a,8f#,8f#,8f#,2g,8p,8p,8f#,8f#,8f#,8g,a#.,8c6,8c6,8c6,c6",
+
+	"MissionImp:d=4,o=6,b=150:16d5,16d#5,16d5,16d#5,16d5,16d#5,16d5,16d5,16d#5,16e5,16f5,16f#5,16g5,8g5,4p,8g5,4p,8a#5,8p,8c6,8p,8g5,4p,8g5,4p,8f5,8p,8p,8g5,4p,4p,8a#5,8p,8c6,8p,8g5,4p,4p,8f5,8p,8f#5,8p,8a#5,8g5,1d5",
+
+	"Batman:d=8,o=5,b=160:16a,16g#,16g,16f#,16f,16f#,16g,16g#,4a.,p,d,d,c#,c#,c,c,c#,c#,d,d,c#,c#,c,c,c#,c#,d,d,c#,c#,c,c,c#,c#,g6,p,4g6",
+
+	"Pinkpanther:d=16,o=5,b=160:8d#,8e,2p,8f#,8g,2p,8d#,8e,p,8f#,8g,p,8c6,8b,p,8d#,8e,p,8b,2a#,2p,a,g,e,d,2e",
+
+	"Adamsfamily:d=4,o=5,b=160:8c,f,8a,f,8c,b4,2g,8f,e,8g,e,8e4,a4,2f,8c,f,8a,f,8c,b4,2g,8f,e,8c,d,8e,1f,8c,8d,8e,8f,1p,8d,8e,8f#,8g,1p,8d,8e,8f#,8g,p,8d,8e,8f#,8g,p,8c,8d,8e,8f",
+
+	"Argentina:d=4,o=5,b=70:8e.4,8e4,8e4,8e.4,8f4,8g4,8a4,g4,8p,8g4,8a4,8a4,8g4,c,g4,8f4,e.4,8p,8e4,8f4,8g4,8d4,d4,8d4,8e4,8f4,c4,16p,8c4,8d4,8c4,8e4,g4,16p,8g4,8g4,8a4,c,16p",
+
+	"Xfiles:d=4,o=5,b=125:e,b,a,b,d6,2b.,1p,e,b,a,b,e6,2b.,1p,g6,f#6,e6,d6,e6,2b.,1p,g6,f#6,e6,d6,f#6,2b.,1p,e,b,a,b,d6,2b.,1p,e,b,a,b,e6,2b.,1p,e6,2b.",
+
+	"Zorba2:d=4,o=5,b=125:16c#6,2d6,2p,16c#6,2d6,2p,32e6,32d6,32c#6,2d6,2p,16c#6,2d6,2p,16b,2c6,2p,32d6,32c6,32b,2c6,2p,16a#,2b,p,8p,32c6,32b,32a,32g,32b,2a,2p,32a,32g,32f#,32a,1g,1p,8c#6,8d6,8d6,8d6,8d6,8d6,8d6,8d6,8c#6,8d6,8d6,8d6,8d6,8d6,16e6,16d6,16c#6,16e6,8c#6,8d6,8d6,8d6,8d6,8d6,8d6,8d6,8c#6,8d6,8d6,8d6,8d6,8d6"
+};
+
+
+// La siguiente matriz almacena las frecuencias de las notas musicales
+const unsigned int note[4][12] =
+{   // C    C#    D     D#    E     F     F#    G     G#    A     A#    B
+	{262, 277, 294, 311, 330, 349, 370, 392, 415, 440, 466, 494},// 4ta octava
+	{523, 554, 587, 622, 659, 698, 740, 784, 830, 880, 932, 988},// 5ta octava
+	{1047,1109,1175,1244,1319,1397,1480,1568,1660,1760,1865,1976},// 6ta octava
+	{2093,2218,2349,2489,2637,2794,2960,3136,3320,3520,3728,3951}  // 7ma octava
+};
+
+void sound(unsigned int, unsigned int);
+void play_song(char*);
+int tocar(void);
+
+#endif /* RINGTONE_H_ */
